updated to Raspberry Pi OS "bookworm" and updated all Greenbone components to match openvas-scanner 22.7.9

+++ scripts/firstrun.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 #
-# Copyright (c) 2023, The beep-projects contributors
+# Copyright (c) 2023-2024, The beep-projects contributors
 # this file originated from https://github.com/beep-projects
 # Do not remove the lines above.
 # This program is free software: you can redistribute it and/or modify
@@ -32,6 +32,7 @@ trap 'exec 2>&4 1>&3' 0 1 2 3
 exec 1>/boot/firstrun.log 2>&1
 
 echo "START firstrun.sh"
+echo "This script is running as user: $( whoami )"
 
 #-------------------------------------------------------------------------------
 #----------------------- START OF CONFIGURATION --------------------------------
@@ -68,28 +69,46 @@ XKBOPTIONS=""
 
 #---- things moved from secondrun.sh ------------------------------------------
 # configure youre locale
+sudo locale-gen "$LOCALE"
 sudo update-locale LANG="$LOCALE"
 # things that require user $USERNAME to exist are move to the end of the script
 #-------------------------------------------------------------------------------
 
+# Prior to Bookworm, Raspberry Pi OS stored the boot partition at /boot/.
+# Since Bookworm, the boot partition is located at /boot/firmware/. 
+echo "checking if files were moved to /boot/firmware"
+if [ -f /boot/firmware/secondrun.sh ]; then
+   echo "ln -s /boot/firmware/secondrun.sh /boot/secondrun.sh"
+   ln -s /boot/firmware/secondrun.sh /boot/secondrun.sh
+fi
+if [ -f /boot/firmware/thirdrun.sh ]; then
+   echo "ln -s /boot/firmware/thirdrun.sh /boot/thirdrun.sh"
+   ln -s /boot/firmware/thirdrun.sh /boot/thirdrun.sh
+fi
 # copy the USERNAME into secondrun.sh 
 sed -i "s/^USERNAME=.*/USERNAME=${USERNAME}/" /boot/secondrun.sh
 
 # set hostname and username
 CURRENT_HOSTNAME=$( </etc/hostname tr -d " \t\n\r" )
 echo "set hostname to ${HOSTNAME} (was ${CURRENT_HOSTNAME})"
-echo $HOSTNAME >/etc/hostname
-sed -i "s/127.0.1.1.*$CURRENT_HOSTNAME/127.0.1.1\t$HOSTNAME/g" /etc/hosts
+if [ -f /usr/lib/raspberrypi-sys-mods/imager_custom ]; then
+   /usr/lib/raspberrypi-sys-mods/imager_custom set_hostname ${HOSTNAME}
+else
+   echo ${HOSTNAME} >/etc/hostname
+   sed -i "s/127.0.1.1.*${CURRENT_HOSTNAME}/127.0.1.1\t${HOSTNAME}/g" /etc/hosts
+fi
 
 FIRSTUSER=$( getent passwd 1000 | cut -d: -f1 )
 echo "set default user to ${USERNAME} (was ${FIRSTUSER})"
+#FIRSTUSERHOME=$( getent passwd 1000 | cut -d: -f6 )
 if [ -f /usr/lib/userconf-pi/userconf ]; then
-   echo "/usr/lib/userconf-pi/userconf ${USERNAME} ${PASSWD}"
+   echo "/usr/lib/userconf-pi/userconf \"${USERNAME}\" \"${PASSWD}\""
    /usr/lib/userconf-pi/userconf "${USERNAME}" "${PASSWD}"
 else
-   echo "setting ${USERNAME}:${PASSWD} the non-Pi-way"
+   echo "echo ${FIRSTUSER}:${PASSWD} | chpasswd -e"
    echo "${FIRSTUSER}:${PASSWD}" | chpasswd -e
    if [ "${FIRSTUSER}" != "${USERNAME}" ]; then
+      echo "user name has changed to: ${USERNAME}"
       usermod -l "${USERNAME}" "${FIRSTUSER}"
       usermod -m -d "/home/${USERNAME}" "${USERNAME}"
       groupmod -n "${USERNAME}" "${FIRSTUSER}"
@@ -106,13 +125,22 @@ else
 fi
 
 echo "setting network options"
-sed -i "s/^REGDOMAIN=.*/REGDOMAIN=${COUNTRY}/" /etc/default/crda
+#sed -i "s/^REGDOMAIN=.*/REGDOMAIN=${COUNTRY}/" /etc/default/crda
 
-systemctl enable ssh
+if [ -f /usr/lib/raspberrypi-sys-mods/imager_custom ]; then
+   /usr/lib/raspberrypi-sys-mods/imager_custom enable_ssh
+else
+   systemctl enable ssh
+fi
+
+if [ -f /usr/lib/raspberrypi-sys-mods/imager_custom ]; then
+   /usr/lib/raspberrypi-sys-mods/imager_custom set_wlan "${SSID}" "${WPA_PASSPHRASE}" "${COUNTRY}"
+else
 cat >/etc/wpa_supplicant/wpa_supplicant.conf <<WPAEOF
-ctrl_interface=DIR=/var/run/wpa_supplicant GROUP=netdev
 country=$COUNTRY
-#ap_scan=1
+ctrl_interface=DIR=/var/run/wpa_supplicant GROUP=netdev
+ap_scan=1
+
 update_config=1
 network={
 	ssid="$SSID"
@@ -120,24 +148,33 @@ network={
 }
 
 WPAEOF
-chmod 600 /etc/wpa_supplicant/wpa_supplicant.conf
-rfkill unblock wifi
-for filename in /var/lib/systemd/rfkill/*:wlan ; do
-  echo 0 > "${filename}"
-done
-rm -f /etc/xdg/autostart/piwiz.desktop
-rm -f /etc/localtime
+   chmod 600 /etc/wpa_supplicant/wpa_supplicant.conf
+   rfkill unblock wifi
+   for filename in /var/lib/systemd/rfkill/*:wlan ; do
+     echo 0 > "${filename}"
+   done
+fi
 
-echo "setting timezone and keyboard layout"
-echo $TIMEZONE >/etc/timezone
-dpkg-reconfigure -f noninteractive tzdata
+#Disable "Welcome to Raspberry Pi" setup wizard at system start
+if [ -f /etc/xdg/autostart/piwiz.desktop ]; then
+   rm -f /etc/xdg/autostart/piwiz.desktop
+fi
+
+if [ -f /usr/lib/raspberrypi-sys-mods/imager_custom ]; then
+   /usr/lib/raspberrypi-sys-mods/imager_custom set_keymap "${COUNTRY}"
+   /usr/lib/raspberrypi-sys-mods/imager_custom set_timezone "${TIMEZONE}"
+else
+   rm -f /etc/localtime
+   echo "${TIMEZONE}" >/etc/timezone
+   dpkg-reconfigure -f noninteractive tzdata
 cat >/etc/default/keyboard <<KBEOF
 XKBMODEL=$XKBMODEL
 XKBLAYOUT=$XKBLAYOUT
 XKBVARIANT=$XKBVARIANT
 XKBOPTIONS=$XKBOPTIONS
 KBEOF
-dpkg-reconfigure -f noninteractive keyboard-configuration
+   dpkg-reconfigure -f noninteractive keyboard-configuration
+fi
 
 #---- things moved from secondrun.sh ------------------------------------------
 # Creating a gvm system user and group
@@ -146,15 +183,11 @@ sudo useradd -r -M -U -G sudo -s /usr/sbin/nologin gvm
 sudo usermod -aG gvm "$USERNAME"
 #------------------------------------------------------------------------------
 
-#clean up
-#echo "removing firstrun.sh from the system"
-#rm -f /boot/firstrun.sh
-sed -i "s| systemd.run.*||g" /boot/cmdline.txt
-
 echo "installing secondrun.service"
 # make sure secondrun.sh is executed at next boot. 
 # we will need network up and running, so we install the script as a service that depends on network
-cat <<EOF >/etc/systemd/system/secondrun.service
+echo "create /etc/systemd/system/secondrun.service"
+cat <<EOF | sudo tee /etc/systemd/system/secondrun.service
 [Unit]
 Description=SecondRun
 After=network.target
@@ -172,10 +205,17 @@ RemainAfterExit=no
 WantedBy=multi-user.target
 
 EOF
+echo "enable secondrun.service"
 #reload systemd to make the daemon aware of the new configuration
 systemctl --system daemon-reload
 #enable service
 systemctl enable secondrun.service
+
+#clean up
+echo "removing firstrun.sh from /boot/cmdline.txt"
+#rm -f /boot/firstrun.sh
+sed -i "s| systemd.run.*||g" /boot/cmdline.txt
+
 echo "DONE firstrun.sh"
 
 exit 0
